Move `getFilterChanges` tests to use HTTP server
