+++ core/lib/zksync_core/src/api_server/web3/tests/mod.rs
@@ -1,3 +1,5 @@
+use assert_matches::assert_matches;
+use async_trait::async_trait;
 use tokio::sync::watch;
 
 use std::{sync::Arc, time::Instant};
@@ -7,13 +9,18 @@ use zksync_config::configs::{
     chain::{NetworkConfig, StateKeeperConfig},
     ContractsConfig,
 };
-use zksync_dal::ConnectionPool;
+use zksync_contracts::BaseSystemContractsHashes;
+use zksync_dal::{transactions_dal::L2TxSubmissionResult, ConnectionPool};
 use zksync_health_check::CheckHealth;
 use zksync_state::PostgresStorageCaches;
-use zksync_types::{L1BatchNumber, U64};
+use zksync_types::{
+    block::MiniblockHeader, fee::TransactionExecutionMetrics, tx::IncludedTxLocation, Address,
+    L1BatchNumber, ProtocolVersionId, VmEvent, H256, U64,
+};
 use zksync_web3_decl::{
-    jsonrpsee::http_client::HttpClient,
+    jsonrpsee::{core::Error as RpcError, http_client::HttpClient, types::error::ErrorCode},
     namespaces::{EthNamespaceClient, ZksNamespaceClient},
+    types::FilterChanges,
 };
 
 mod ws;
@@ -133,8 +140,12 @@ async fn spawn_server(
     (server_handles, pub_sub_events_receiver)
 }
 
-#[tokio::test]
-async fn http_server_can_start() {
+#[async_trait]
+trait HttpTest {
+    async fn test(&self, client: &HttpClient, pool: &ConnectionPool) -> anyhow::Result<()>;
+}
+
+async fn test_http_server(test: impl HttpTest) {
     let pool = ConnectionPool::test_pool().await;
     let network_config = NetworkConfig::for_tests();
     let mut storage = pool.access_storage().await.unwrap();
@@ -150,29 +161,343 @@ async fn http_server_can_start() {
     drop(storage);
 
     let (stop_sender, stop_receiver) = watch::channel(false);
-    let server_handles = spawn_http_server(&network_config, pool, stop_receiver).await;
+    let server_handles = spawn_http_server(&network_config, pool.clone(), stop_receiver).await;
     server_handles.wait_until_ready().await;
 
-    test_http_server_methods(server_handles.local_addr).await;
+    let client = <HttpClient>::builder()
+        .build(format!("http://{}/", server_handles.local_addr))
+        .unwrap();
+    test.test(&client, &pool).await.unwrap();
 
     stop_sender.send_replace(true);
     server_handles.shutdown().await;
 }
 
-async fn test_http_server_methods(local_addr: SocketAddr) {
-    let client = <HttpClient>::builder()
-        .build(format!("http://{local_addr}/"))
-        .unwrap();
-    let block_number = client.get_block_number().await.unwrap();
-    assert_eq!(block_number, U64::from(0));
+fn assert_logs_match(actual_logs: &[api::Log], expected_logs: &[&VmEvent]) {
+    assert_eq!(actual_logs.len(), expected_logs.len());
+    for (actual_log, &expected_log) in actual_logs.iter().zip(expected_logs) {
+        assert_eq!(actual_log.address, expected_log.address);
+        assert_eq!(actual_log.topics, expected_log.indexed_topics);
+        assert_eq!(actual_log.data.0, expected_log.value);
+    }
+}
 
-    let l1_batch_number = client.get_l1_batch_number().await.unwrap();
-    assert_eq!(l1_batch_number, U64::from(0));
+fn create_miniblock(number: u32) -> MiniblockHeader {
+    MiniblockHeader {
+        number: MiniblockNumber(number),
+        timestamp: number.into(),
+        hash: H256::from_low_u64_be(number.into()),
+        l1_tx_count: 0,
+        l2_tx_count: 0,
+        base_fee_per_gas: 100,
+        l1_gas_price: 100,
+        l2_fair_gas_price: 100,
+        base_system_contracts_hashes: BaseSystemContractsHashes::default(),
+        protocol_version: Some(ProtocolVersionId::latest()),
+        virtual_blocks: 1,
+    }
+}
 
-    let genesis_l1_batch = client
-        .get_l1_batch_details(L1BatchNumber(0))
-        .await
-        .unwrap()
-        .unwrap();
-    assert!(genesis_l1_batch.base.root_hash.is_some());
+async fn store_block(pool: &ConnectionPool) -> anyhow::Result<(MiniblockHeader, H256)> {
+    let mut storage = pool.access_storage().await?;
+    let new_tx = create_l2_transaction(1, 2);
+    let new_tx_hash = new_tx.hash();
+    let tx_submission_result = storage
+        .transactions_dal()
+        .insert_transaction_l2(new_tx, TransactionExecutionMetrics::default())
+        .await;
+    assert_matches!(tx_submission_result, L2TxSubmissionResult::Added);
+
+    let new_miniblock = create_miniblock(1);
+    storage
+        .blocks_dal()
+        .insert_miniblock(&new_miniblock)
+        .await?;
+    Ok((new_miniblock, new_tx_hash))
+}
+
+async fn store_events(
+    storage: &mut StorageProcessor<'_>,
+    miniblock_number: u32,
+    start_idx: u32,
+) -> anyhow::Result<(IncludedTxLocation, Vec<VmEvent>)> {
+    let new_miniblock = create_miniblock(miniblock_number);
+    storage
+        .blocks_dal()
+        .insert_miniblock(&new_miniblock)
+        .await?;
+    let tx_location = IncludedTxLocation {
+        tx_hash: H256::repeat_byte(1),
+        tx_index_in_miniblock: 0,
+        tx_initiator_address: Address::repeat_byte(2),
+    };
+    let events = vec![
+        // Matches address, doesn't match topics
+        VmEvent {
+            location: (L1BatchNumber(1), start_idx),
+            address: Address::repeat_byte(23),
+            indexed_topics: vec![],
+            value: start_idx.to_le_bytes().to_vec(),
+        },
+        // Doesn't match address, matches topics
+        VmEvent {
+            location: (L1BatchNumber(1), start_idx + 1),
+            address: Address::zero(),
+            indexed_topics: vec![H256::repeat_byte(42)],
+            value: (start_idx + 1).to_le_bytes().to_vec(),
+        },
+        // Doesn't match address or topics
+        VmEvent {
+            location: (L1BatchNumber(1), start_idx + 2),
+            address: Address::zero(),
+            indexed_topics: vec![H256::repeat_byte(1), H256::repeat_byte(42)],
+            value: (start_idx + 2).to_le_bytes().to_vec(),
+        },
+        // Matches both address and topics
+        VmEvent {
+            location: (L1BatchNumber(1), start_idx + 3),
+            address: Address::repeat_byte(23),
+            indexed_topics: vec![H256::repeat_byte(42), H256::repeat_byte(111)],
+            value: (start_idx + 3).to_le_bytes().to_vec(),
+        },
+    ];
+    storage
+        .events_dal()
+        .save_events(
+            MiniblockNumber(miniblock_number),
+            &[(tx_location, events.iter().collect())],
+        )
+        .await;
+    Ok((tx_location, events))
+}
+
+#[derive(Debug)]
+struct HttpServerBasics;
+
+#[async_trait]
+impl HttpTest for HttpServerBasics {
+    async fn test(&self, client: &HttpClient, _pool: &ConnectionPool) -> anyhow::Result<()> {
+        let block_number = client.get_block_number().await?;
+        assert_eq!(block_number, U64::from(0));
+
+        let l1_batch_number = client.get_l1_batch_number().await?;
+        assert_eq!(l1_batch_number, U64::from(0));
+
+        let genesis_l1_batch = client
+            .get_l1_batch_details(L1BatchNumber(0))
+            .await?
+            .context("No genesis L1 batch")?;
+        assert!(genesis_l1_batch.base.root_hash.is_some());
+        Ok(())
+    }
+}
+
+#[tokio::test]
+async fn http_server_basics() {
+    test_http_server(HttpServerBasics).await;
+}
+
+#[derive(Debug)]
+struct BasicFilterChanges;
+
+#[async_trait]
+impl HttpTest for BasicFilterChanges {
+    async fn test(&self, client: &HttpClient, pool: &ConnectionPool) -> anyhow::Result<()> {
+        let block_filter_id = client.new_block_filter().await?;
+        let tx_filter_id = client.new_pending_transaction_filter().await?;
+
+        let (new_miniblock, new_tx_hash) = store_block(pool).await?;
+
+        let block_filter_changes = client.get_filter_changes(block_filter_id).await?;
+        assert_matches!(
+            block_filter_changes,
+            FilterChanges::Hashes(hashes) if hashes == [new_miniblock.hash]
+        );
+        let block_filter_changes = client.get_filter_changes(block_filter_id).await?;
+        assert_matches!(block_filter_changes, FilterChanges::Hashes(hashes) if hashes.is_empty());
+
+        let tx_filter_changes = client.get_filter_changes(tx_filter_id).await?;
+        assert_matches!(
+            tx_filter_changes,
+            FilterChanges::Hashes(hashes) if hashes == [new_tx_hash]
+        );
+        let tx_filter_changes = client.get_filter_changes(tx_filter_id).await?;
+        assert_matches!(tx_filter_changes, FilterChanges::Hashes(hashes) if hashes.is_empty());
+
+        // Check uninstalling the filter.
+        let removed = client.uninstall_filter(block_filter_id).await?;
+        assert!(removed);
+        let removed = client.uninstall_filter(block_filter_id).await?;
+        assert!(!removed);
+
+        let err = client
+            .get_filter_changes(block_filter_id)
+            .await
+            .unwrap_err();
+        assert_matches!(err, RpcError::Call(err) if err.code() == ErrorCode::InvalidParams.code());
+        Ok(())
+    }
+}
+
+#[tokio::test]
+async fn basic_filter_changes() {
+    test_http_server(BasicFilterChanges).await;
+}
+
+#[derive(Debug)]
+struct LogFilterChanges;
+
+#[async_trait]
+impl HttpTest for LogFilterChanges {
+    async fn test(&self, client: &HttpClient, pool: &ConnectionPool) -> anyhow::Result<()> {
+        let all_logs_filter_id = client.new_filter(Filter::default()).await?;
+        let address_filter = Filter {
+            address: Some(Address::repeat_byte(23).into()),
+            ..Filter::default()
+        };
+        let address_filter_id = client.new_filter(address_filter).await?;
+        let topics_filter = Filter {
+            topics: Some(vec![Some(H256::repeat_byte(42).into())]),
+            ..Filter::default()
+        };
+        let topics_filter_id = client.new_filter(topics_filter).await?;
+
+        let mut storage = pool.access_storage().await?;
+        let (_, events) = store_events(&mut storage, 1, 0).await?;
+        drop(storage);
+        let events: Vec<_> = events.iter().collect();
+
+        let all_logs = client.get_filter_changes(all_logs_filter_id).await?;
+        let FilterChanges::Logs(all_logs) = all_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", all_logs);
+        };
+        assert_logs_match(&all_logs, &events);
+
+        let address_logs = client.get_filter_changes(address_filter_id).await?;
+        let FilterChanges::Logs(address_logs) = address_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", address_logs);
+        };
+        assert_logs_match(&address_logs, &[events[0], events[3]]);
+
+        let topics_logs = client.get_filter_changes(topics_filter_id).await?;
+        let FilterChanges::Logs(topics_logs) = topics_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", topics_logs);
+        };
+        assert_logs_match(&topics_logs, &[events[1], events[3]]);
+
+        let new_all_logs = client.get_filter_changes(all_logs_filter_id).await?;
+        let FilterChanges::Logs(new_all_logs) = new_all_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", new_all_logs);
+        };
+        assert_eq!(new_all_logs, all_logs); // FIXME: is this expected?
+        Ok(())
+    }
+}
+
+#[tokio::test]
+async fn log_filter_changes() {
+    test_http_server(LogFilterChanges).await;
+}
+
+#[derive(Debug)]
+struct LogFilterChangesWithBlockBoundaries;
+
+#[async_trait]
+impl HttpTest for LogFilterChangesWithBlockBoundaries {
+    async fn test(&self, client: &HttpClient, pool: &ConnectionPool) -> anyhow::Result<()> {
+        let lower_bound_filter = Filter {
+            from_block: Some(api::BlockNumber::Number(2.into())),
+            ..Filter::default()
+        };
+        let lower_bound_filter_id = client.new_filter(lower_bound_filter).await?;
+        let upper_bound_filter = Filter {
+            to_block: Some(api::BlockNumber::Number(1.into())),
+            ..Filter::default()
+        };
+        let upper_bound_filter_id = client.new_filter(upper_bound_filter).await?;
+        let bounded_filter = Filter {
+            from_block: Some(api::BlockNumber::Number(1.into())),
+            to_block: Some(api::BlockNumber::Number(1.into())),
+            ..Filter::default()
+        };
+        let bounded_filter_id = client.new_filter(bounded_filter).await?;
+
+        let mut storage = pool.access_storage().await?;
+        let (_, events) = store_events(&mut storage, 1, 0).await?;
+        drop(storage);
+        let events: Vec<_> = events.iter().collect();
+
+        let lower_bound_logs = client.get_filter_changes(lower_bound_filter_id).await?;
+        assert_matches!(
+            lower_bound_logs,
+            FilterChanges::Hashes(hashes) if hashes.is_empty()
+        );
+        // ^ Since `FilterChanges` is serialized w/o a tag, an empty array will be deserialized
+        // as `Hashes(_)` (the first declared variant).
+
+        let upper_bound_logs = client.get_filter_changes(upper_bound_filter_id).await?;
+        let FilterChanges::Logs(upper_bound_logs) = upper_bound_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", upper_bound_logs);
+        };
+        assert_logs_match(&upper_bound_logs, &events);
+        let bounded_logs = client.get_filter_changes(bounded_filter_id).await?;
+        let FilterChanges::Logs(bounded_logs) = bounded_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", bounded_logs);
+        };
+        assert_eq!(bounded_logs, upper_bound_logs);
+
+        // Add another miniblock with events to the storage.
+        let mut storage = pool.access_storage().await?;
+        let (_, new_events) = store_events(&mut storage, 2, 4).await?;
+        drop(storage);
+        let new_events: Vec<_> = new_events.iter().collect();
+
+        let lower_bound_logs = client.get_filter_changes(lower_bound_filter_id).await?;
+        let FilterChanges::Logs(lower_bound_logs) = lower_bound_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", lower_bound_logs);
+        };
+        assert_logs_match(&lower_bound_logs, &new_events);
+
+        // FIXME: is this expected?
+        let new_upper_bound_logs = client.get_filter_changes(upper_bound_filter_id).await?;
+        assert_eq!(new_upper_bound_logs, FilterChanges::Logs(upper_bound_logs));
+        let new_bounded_logs = client.get_filter_changes(bounded_filter_id).await?;
+        assert_eq!(new_bounded_logs, FilterChanges::Logs(bounded_logs));
+
+        // Add miniblock #3. It should not be picked up by the bounded and upper bound filters,
+        // and should be picked up by the lower bound filter.
+        let mut storage = pool.access_storage().await?;
+        let (_, new_events) = store_events(&mut storage, 3, 8).await?;
+        drop(storage);
+        let new_events: Vec<_> = new_events.iter().collect();
+
+        let bounded_logs = client.get_filter_changes(bounded_filter_id).await?;
+        let FilterChanges::Logs(bounded_logs) = bounded_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", bounded_logs);
+        };
+        assert!(bounded_logs
+            .iter()
+            .all(|log| log.block_number.unwrap() < 3.into()));
+
+        let upper_bound_logs = client.get_filter_changes(upper_bound_filter_id).await?;
+        let FilterChanges::Logs(upper_bound_logs) = upper_bound_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", upper_bound_logs);
+        };
+        assert!(upper_bound_logs
+            .iter()
+            .all(|log| log.block_number.unwrap() < 3.into()));
+
+        let lower_bound_logs = client.get_filter_changes(lower_bound_filter_id).await?;
+        let FilterChanges::Logs(lower_bound_logs) = lower_bound_logs else {
+            panic!("Unexpected getFilterChanges output: {:?}", lower_bound_logs);
+        };
+        let start_idx = lower_bound_logs.len() - 4;
+        assert_logs_match(&lower_bound_logs[start_idx..], &new_events);
+        Ok(())
+    }
+}
+
+#[tokio::test]
+async fn log_filter_changes_with_block_boundaries() {
+    test_http_server(LogFilterChangesWithBlockBoundaries).await;
 }
